+++ R/which_can_be.R
@@ -1,3 +1,11 @@
+#' @details 
+#' 
+#' `which_can_be` provides indices of `object`` that **can** be coerced to the given 
+#' `class`.  `NA` can always be coerced. 
+#' 
+#' `which_cant_be` provides indices of `object` that **can't** be coerced to the 
+#' given `class`. `NA` can always be coerced.
+#' 
 #' @return 
 #' `which_can_be` returns a integer vector of the indices of values that can be
 #' successfully coerced to  the new class.
@@ -46,18 +54,41 @@
 #'   posx %>% which_can_be("Date")
 #'   posx %>% which_can_be("POSIXct") 
 #'   
+#'   
+#'   x <- c( 1:3, letters[1:3], NA )
+#'   x %>% which_can_be( "integer" )
+#'   x %>% which_cant_be( "integer" )
+#'   
 #' @md
 #' @rdname can_be
 #' @export
 
 which_can_be <- function( object, class ) { 
   
-  which( 
-    is.na(object)  |   # Existing NA values in object
-    ! is.na( try_as(object, class) )  # Values 
-  ) 
+  suppressWarnings(
+    which( 
+      is.na(object)  |   # Existing NA values in object
+      ! is.na( try( as(object, class), silent = TRUE ) )  # Values 
+    ) 
+  )
   
 }
+
+#' @md
+#' @rdname can_be
+#' @export
+
+which_cant_be <- function( object, class ) { 
+
+  suppressWarnings(
+    which( 
+     ( ! is.na(object) ) &   # Existing NA values in object
+     is.na( try( as( object, class), silent = TRUE ) )  # Values 
+    ) 
+  )
+  
+}
+
 
 
 #' @rdname can_be