+++ NEWS.md
@@ -1,9 +1,11 @@
 # Version 0.3.0 
 
  - Refactor `can_be` code:
-   - Begin Deprecation of `can`
-   - Add `try_as`
-   - Add `which_can_be` 
+   - Begin Deprecation of `can` 
+ - Add `try_as`
+ - Add `which_can_be` and `which_cant_be` and tests
+ - Improved package start-up
+ - Added markdown for man
 
 # Version 0.2.5